pyproject.toml
[project]
name = "cb-assit-langchain"
version = "0.2.0"
description = "AI-powered coloring book assistant using LangGraph multi-agent system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.32",
    "langchain-core>=0.3.0",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "openai>=1.0.0",
    "duckduckgo-search>=6.0.0",
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
]

[project.scripts]
midjourney-agent = "integrations.midjourney.cli:main"
